[finance_report_vision] Split ai_advisor.py (1089 lines) into a package
Behaviour-preserving split of the AI advisor module into a ai_advisor/ package;
public import surface preserved. AIAdvisorService is kept whole (under 1000
lines); only the module-level helpers are extracted.

- ai_advisor/_base.py: shared constants/patterns + logger.
- ai_advisor/_guardrails.py: language/injection/sensitive/non-financial checks,
  redaction, disclaimer, StreamRedactor.
- ai_advisor/_cache.py: ResponseCache + the _CACHE singleton.
- ai_advisor/service.py: AIAdvisorError, ChatStream, AIAdvisorService.
- __init__: re-exports the full public surface (incl. guardrail fns, ResponseCache,
  _CACHE, PortfolioService) so tests need no import changes for those.

Test impact (test-only): the module-alias used for monkeypatching is repointed
to the `service` submodule (where the patched names are bound). The AST
transaction-boundary guard now scans service packages recursively (rglob +
relative-path keys) so the _stream_and_store commit boundary stays guarded.
117 ai_advisor tests pass; full collection clean.
