+++ apps/backend/src/services/ai_advisor/service.py
@@ -1,4 +1,4 @@
-"""AI advisor service for conversational financial insights."""
+"""AIAdvisorService + chat stream/error types."""
 
 from __future__ import annotations
 
@@ -6,7 +6,6 @@ import asyncio
 import hashlib
 import json
 import re
-import time
 from collections.abc import AsyncIterator
 from dataclasses import dataclass, field
 from datetime import UTC, date, datetime
@@ -21,7 +20,6 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from src.config import settings
 from src.llm.common import ReasoningEffort, Scene, SceneBinding
 from src.llm.factory import get_config_source
-from src.logger import get_logger
 from src.models import (
     AccountType,
     ChatMessage,
@@ -30,8 +28,23 @@ from src.models import (
     ChatSessionStatus,
 )
 from src.money import to_money
-from src.prompts.ai_advisor import DISCLAIMER_EN, DISCLAIMER_ZH, get_ai_advisor_prompt
+from src.prompts.ai_advisor import get_ai_advisor_prompt
 from src.schemas.chat import AdvisorSuggestion, ChatActionChip, ChatCitation, ChatResponseMetadata
+from src.services.ai_advisor._base import CHAT_METADATA_SAFE_HREFS, CONFIDENCE_WORST_ORDER, MAX_CONTEXT_MESSAGES, logger
+from src.services.ai_advisor._cache import _CACHE
+from src.services.ai_advisor._guardrails import (
+    StreamRedactor,
+    build_refusal,
+    detect_language,
+    ensure_disclaimer,
+    estimate_tokens,
+    is_non_financial,
+    is_prompt_injection,
+    is_sensitive_request,
+    is_write_request,
+    normalize_question,
+    redact_sensitive,
+)
 from src.services.ai_streaming import stream_ai_chat
 from src.services.market_data import MarketDataScopeStatus, get_market_data_status
 from src.services.portfolio import PortfolioNotFoundError, PortfolioService
@@ -44,115 +57,6 @@ from src.services.reporting import (
     get_category_breakdown,
 )
 from src.services.workflow_events import get_workflow_status
-
-logger = get_logger(__name__)
-
-MAX_CONTEXT_MESSAGES = 20
-CACHE_TTL_SECONDS = 3600
-
-INJECTION_PATTERNS = (
-    r"ignore (all|previous|prior) instructions",
-    r"disregard (all|previous|prior) instructions",
-    r"reveal (the )?(system|developer) prompt",
-    r"system prompt",
-    r"developer message",
-    r"jailbreak",
-    r"bypass safety",
-    r"override (rules|policy)",
-    r"forget what we talked about",
-    r"you are now a",
-)
-
-SENSITIVE_PATTERNS = (
-    r"password",
-    r"account number",
-    r"card number",
-    r"credit card",
-    r"cvv",
-    r"otp",
-    r"pin",
-    r"social security",
-    r"ssn",
-    r"secret key",
-)
-
-NON_FINANCIAL_PATTERNS = (
-    r"weather",
-    r"joke",
-    r"movie",
-    r"music",
-    r"recipe",
-    r"sports",
-    r"news",
-    r"code",
-    r"programming",
-)
-
-CHAT_METADATA_SAFE_HREFS = (
-    "/reports/balance-sheet",
-    "/reports/income-statement",
-    "/reports/package",
-    "/reports",
-    "/reconciliation/review-queue",
-    "/review",
-    "/portfolio/prices",
-    "/portfolio",
-    "/assets",
-    "/statements/upload",
-)
-
-CONFIDENCE_WORST_ORDER = {
-    "LOW": 0,
-    "MEDIUM": 1,
-    "HIGH": 2,
-    "TRUSTED": 3,
-    "DETERMINISTIC": 3,
-}
-
-# SECURITY: Match long number sequences but avoid common date/time patterns
-# (YYYY-MM-DD, DD/MM/YYYY, etc.)
-# Generic branch: 12+ consecutive digits (no separators) to reduce accidental date matches
-SENSITIVE_NUMBER_RE = re.compile(
-    r"(?<!\d)\d{12,}(?!\d)"  # 12+ contiguous digits (account numbers, identifiers)
-    r"|"
-    r"(?<!\d)(?:\d{4}[ -]?){3}\d{4}(?!\d)"  # Credit card format: 4x4 digits
-)
-
-DISCLAIMER_BY_LANG = {
-    "en": DISCLAIMER_EN,
-    "zh": DISCLAIMER_ZH,
-}
-
-REFUSAL_BY_REASON = {
-    "injection": {
-        "en": "I cannot help with that request. Please ask a finance-related question.",
-        "zh": (
-            "\u6211\u65e0\u6cd5\u534f\u52a9\u8be5\u8bf7\u6c42\u3002"
-            "\u8bf7\u63d0\u51fa\u4e0e\u8d22\u52a1\u76f8\u5173\u7684\u95ee\u9898\u3002"
-        ),
-    },
-    "write": {
-        "en": (
-            "I can only provide read-only analysis and cannot create or modify ledger entries. "
-            "Please use the manual entry workflow."
-        ),
-        "zh": (
-            "\u6211\u53ea\u80fd\u63d0\u4f9b\u53ea\u8bfb\u5206\u6790\uff0c"
-            "\u65e0\u6cd5\u521b\u5efa\u6216\u4fee\u6539\u5206\u5f55\u3002"
-            "\u8bf7\u4f7f\u7528\u624b\u52a8\u5f55\u5165\u6d41\u7a0b\u3002"
-        ),
-    },
-    "sensitive": {
-        "en": "For safety reasons, I cannot provide sensitive information.",
-        "zh": (
-            "\u51fa\u4e8e\u5b89\u5168\u539f\u56e0\uff0c\u6211\u65e0\u6cd5\u63d0\u4f9b\u654f\u611f\u4fe1\u606f\u3002"
-        ),
-    },
-    "non_financial": {
-        "en": "This assistant only answers finance-related questions.",
-        "zh": ("\u8be5\u52a9\u624b\u4ec5\u56de\u7b54\u8d22\u52a1\u76f8\u5173\u95ee\u9898\u3002"),
-    },
-}
 
 
 class AIAdvisorError(Exception):
@@ -168,136 +72,6 @@ class ChatStream:
     model_name: str | None
     cached: bool
     metadata: ChatResponseMetadata = field(default_factory=ChatResponseMetadata)
-
-
-class ResponseCache:
-    """Simple in-memory cache for common answers."""
-
-    def __init__(self, ttl_seconds: int = CACHE_TTL_SECONDS) -> None:
-        self._ttl = ttl_seconds
-        self._store: dict[str, tuple[float, str]] = {}
-
-    def get(self, key: str) -> str | None:
-        now = time.time()
-        entry = self._store.get(key)
-        if not entry:
-            return None
-        expires_at, value = entry
-        if now >= expires_at:
-            self._store.pop(key, None)
-            return None
-        return value
-
-    def set(self, key: str, value: str) -> None:
-        expires_at = time.time() + self._ttl
-        self._store[key] = (expires_at, value)
-
-    def prune(self) -> None:
-        now = time.time()
-        expired = [key for key, (exp, _) in self._store.items() if exp <= now]
-        for key in expired:
-            self._store.pop(key, None)
-
-
-_CACHE = ResponseCache()
-
-
-def detect_language(message: str) -> str:
-    """Detect message language based on CJK presence."""
-    if re.search(r"[\u4e00-\u9fff]", message):
-        return "zh"
-    return "en"
-
-
-def normalize_question(message: str) -> str:
-    """Normalize question string for caching."""
-    cleaned = re.sub(r"\s+", " ", message.strip().lower())
-    normalized = re.sub(r"[^a-z0-9\s]", "", cleaned)
-    if normalized:
-        return normalized
-    return hashlib.sha1(message.encode("utf-8")).hexdigest()
-
-
-def estimate_tokens(text: str) -> int:
-    """Estimate token count for usage tracking."""
-    return max(1, len(text) // 4)
-
-
-def _matches_any(message: str, patterns: tuple[str, ...]) -> bool:
-    return any(re.search(pattern, message, re.IGNORECASE) for pattern in patterns)
-
-
-def is_prompt_injection(message: str) -> bool:
-    return _matches_any(message, INJECTION_PATTERNS)
-
-
-def is_sensitive_request(message: str) -> bool:
-    return _matches_any(message, SENSITIVE_PATTERNS)
-
-
-def is_non_financial(message: str) -> bool:
-    return _matches_any(message, NON_FINANCIAL_PATTERNS)
-
-
-def is_write_request(message: str) -> bool:
-    return _matches_any(
-        message,
-        (
-            r"create (a )?(journal|entry)",
-            r"post (a )?(journal|entry)",
-            r"delete (a )?(journal|entry)",
-            r"void (a )?(journal|entry)",
-            r"modify (a )?(ledger|entry)",
-            r"write (a )?(ledger|entry)",
-        ),
-    )
-
-
-def redact_sensitive(text: str) -> str:
-    """Redact long number sequences to avoid sensitive data leaks."""
-    return SENSITIVE_NUMBER_RE.sub("[REDACTED]", text)
-
-
-def ensure_disclaimer(text: str, language: str) -> str:
-    """Ensure the output ends with the disclaimer."""
-    disclaimer = DISCLAIMER_BY_LANG.get(language, DISCLAIMER_EN)
-    if text.strip().endswith(disclaimer):
-        return text
-    if not text.endswith("\n"):
-        text += "\n\n"
-    return f"{text}{disclaimer}"
-
-
-def build_refusal(reason: str, language: str) -> str:
-    """Build refusal response with disclaimer."""
-    message = REFUSAL_BY_REASON.get(reason, REFUSAL_BY_REASON["non_financial"]).get(
-        language, REFUSAL_BY_REASON["non_financial"]["en"]
-    )
-    return ensure_disclaimer(message, language)
-
-
-class StreamRedactor:
-    """Streaming redactor that masks sensitive numbers without breaking chunks."""
-
-    def __init__(self, tail_size: int = 64) -> None:
-        self._tail_size = tail_size
-        self._buffer = ""
-
-    def process(self, chunk: str) -> str:
-        combined = self._buffer + chunk
-        if len(combined) <= self._tail_size:
-            self._buffer = combined
-            return ""
-        safe_part = combined[: -self._tail_size]
-        self._buffer = combined[-self._tail_size :]
-        return redact_sensitive(safe_part)
-
-    def flush(self) -> str:
-        if not self._buffer:
-            return ""
-        output = redact_sensitive(self._buffer)
-        self._buffer = ""
-        return output
 
 
 async def _bound_scene_binding(user_id: UUID | None) -> SceneBinding | None:
